vultisig-analytics/requirements.txt
# requirements.txt
psycopg2-binary>=2.9.9
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
schedule==1.2.0
asyncpg==0.28.0
aiohttp==3.8.6
pandas==2.1.1
sqlalchemy==2.0.21
alembic==1.12.0